pyproject.toml

[tool.poetry]
name = "pyneoncli"
version = "0.0.9a1"
description = "A Python CLI for the Neon API"
authors = ["Joe Drumgoole <Joe.Drumgoole@neon.tech>"]
license = "Apache-2.0"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.8,<4.0"
requests = "^2.30.0"
pygments = "^2.15.1"
poetry = "^1.4.2"
colorama = "^0.4.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[tool.poetry.scripts]
neoncli = "pyneoncli.climain:main"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
addopts = "--cov=my_project --cov-report=html"

